feat(*): rewrite entire lib from scratch
The lib has been entirely rewritten in order to have the source code in TypeScript and leveraging
the "tsdx" package for bundling, testing and linting.

BREAKING CHANGE: As part of the rewrite, the global exports are gone. Please use CommonJS or ESM
style imports as specified in the README.

fixes #55 closes #57