refactor(if): conditionally check if error should be thrown
If NODE_ENV is 'production' then the first part of the evaluation short-circuits the conditional
passed into handleError, ensuring that in production environments the check for whether the error
has to be thrown at all is not resolved, ultimately increasing performance. Furthermore if NODE_ENV
is undefined or an empty string it is also assumed to be production, to provide for those who do not
use a common framework such as create-react-app, NextJS or Gatsby.